SystemUI: fix incorrect tile shifting behavior
We need to clamp the max row count to 3 so that the method which
shifts any given tile to the next position can properly account when to
put a tile on the next page.

Also removed some extra debugging, make sure the user can't end up in a
blank page after removing the last tile.

Change-Id: I9367f35b1970c63383ab07dee2bf0622fcb8e94a
Signed-off-by: Roman Birg <roman@cyngn.com>