KeyguardSecurityViewFlipper : Fix sim unlock height overlap
If the font size is set to Huge, and pin lockscreen + sim unlock screen
is showing, the text for the pinpad overlaps. Reason is we enforce the smallest
maxHeight on all children, and this breaks certain screens. Just let each child
use its own maxHeight that it set to begin with.

FEIJ-184

Change-Id: I3b461f6f75a3f541281fb32efc83a891ca7b65f0